backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Steven Jiang",email = "steve@clarative.ai"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn[standard] (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "openai (>=2.31.0,<3.0.0)",
    "greenlet (>=3.4.0,<4.0.0)",
    "aiodocker (>=0.26.0,<0.27.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
